requirements.txt
discord.py==2.3.2
gdown==5.2.0
Flask==2.3.3
